// src/erp-ui/projects/api/src/lib/model/delivery-dto.ts
/**
 * Fachschaften-ERP.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DeliveryPositionDto } from './delivery-position-dto';


export interface DeliveryDto { 
    positions: Array<DeliveryPositionDto>;
    id: string;
    supplierId: string;
    deliveryDate: string;
    documentNumber?: string | null;
    receiptUrl?: string | null;
    note?: string | null;
}